refactor(tui): colocate Screen::Confirm into tui/screens/confirm.rs
